only register the agent if not already registered

+++ install.sh
@@ -82,6 +82,7 @@ if [ $DISTRIBUTION = "Darwin" ]; then
   OS="macOS"
 elif [ -f /etc/debian_version -o "$DISTRIBUTION" == "Debian" -o "$DISTRIBUTION" == "Ubuntu" ]; then
   OS="Debian"
+  FILE=/etc/opt/mondoo/mondoo.yml
 elif [ -f /etc/redhat-release -o "$DISTRIBUTION" == "RedHat" -o "$DISTRIBUTION" == "CentOS" -o "$DISTRIBUTION" == "Amazon" ]; then
   OS="RedHat"
 # openSUSE and SUSE use /etc/SuSE-release
@@ -142,33 +143,38 @@ fi
 
 # Lets register the agent
 if [ ! -z "${MONDOO_REGISTRATION_TOKEN}" ]; then
+  # if mondoo creds file does not exist, register the token
+  if ! test -f "$FILE"; then
+    if [ $(cat /proc/1/comm) = "init" ]
+    then
+      echo " -> Stop mondoo upstart service"
+      $sudo_cmd stop mondoo || true
+    elif [ $(cat /proc/1/comm) = "systemd" ]
+    then
+      echo " -> Stop mondoo systemd service"
+      $sudo_cmd systemctl stop mondoo
+    fi
 
-  if [ $(cat /proc/1/comm) = "init" ]
-  then
-    echo " -> Stop mondoo upstart service"
-    $sudo_cmd stop mondoo || true
-  elif [ $(cat /proc/1/comm) = "systemd" ]
-  then
-    echo " -> Stop mondoo systemd service"
-    $sudo_cmd systemctl stop mondoo
-  fi
+    purple_bold "\n* Register agent with Mondoo Cloud"
+    $sudo_cmd mkdir -p /etc/opt/mondoo/
+    $sudo_cmd mondoo register --config /etc/opt/mondoo/mondoo.yml --token $MONDOO_REGISTRATION_TOKEN
 
-  purple_bold "\n* Register agent with Mondoo Cloud"
-  $sudo_cmd mkdir -p /etc/opt/mondoo/
-  $sudo_cmd mondoo register --config /etc/opt/mondoo/mondoo.yml --token $MONDOO_REGISTRATION_TOKEN
-
-  if [ $(cat /proc/1/comm) = "init" ]
-  then
-    purple_bold "\n* Configuring upstart service"
-    $sudo_cmd start mondoo || true
-  elif [ $(cat /proc/1/comm) = "systemd" ]
-  then
-    purple_bold "\n* Configuring systemd service"
-    $sudo_cmd systemctl enable mondoo.service
-    $sudo_cmd systemctl start mondoo.service
-    $sudo_cmd systemctl daemon-reload
+    if [ $(cat /proc/1/comm) = "init" ]
+    then
+      purple_bold "\n* Configuring upstart service"
+      $sudo_cmd start mondoo || true
+    elif [ $(cat /proc/1/comm) = "systemd" ]
+    then
+      purple_bold "\n* Configuring systemd service"
+      $sudo_cmd systemctl enable mondoo.service
+      $sudo_cmd systemctl start mondoo.service
+      $sudo_cmd systemctl daemon-reload
+    else
+      red "\nSkip service setup: could not detect a supported init system"
+    fi
   else
-    red "\nSkip service setup: could not detect a supported init system"
+    purple_bold "\n* Agent already registered. Skipping registration."
+    is_registered=true
   fi
 else
   red "\nSkip agent registration since MONDOO_REGISTRATION_TOKEN was not set"
@@ -189,7 +195,12 @@ else
 fi
 
 # Display final message
-purple_bold "\nThank you for installing Mondoo!"
+if [ "$is_registered" = true ]; then
+  action="using"
+else
+  action="installing"
+fi
+purple_bold "\nThank you for $action Mondoo!"
 purple "
 If you have any questions, please reach out at Mondoo Community:
